+++ weady/weady/Sources/Common/Navigation/BaseTabScreen.swift
@@ -7,6 +7,11 @@
 
 import SwiftUI
 
+// MARK: - BaseTabScreen
+/// 현재 선택된 탭에 맞는 플로우 호스트 보여줌
+/// 각 플로우 호스트는 자기만의 NavigationStack을 소유
+/// 플로우별 Route/Router/목적지를 분리
+
 struct BaseTabScreen: View {
     @Binding var selectedTab: TabType
 
@@ -15,17 +20,16 @@ struct BaseTabScreen: View {
             Group {
                 switch selectedTab {
                 case .home:
-                    HomeFlowHost()
+                    HomeFlowHost() // 홈 플로우 스택
                 case .weadyboard:
-                    WeadyboardFlowHost()
+                    WeadyboardFlowHost() // 웨디보드 플로우 스택
                 case .weadychive:
-                    WeadychiveFlowHost()
+                    WeadychiveFlowHost() // 웨디카이브 플로우 스택
                 case .mypage:
-                    MyPageFlowHost()
+                    MyPageFlowHost() // 마이페이지 플로우 스택
                 }
             }
             .frame(maxHeight: .infinity)
-            
         }
         .ignoresSafeArea(edges: .bottom)
     }
