+++ weady/weady/Sources/Common/Navigation/NavigationRouter.swift
@@ -13,7 +13,7 @@ class NavigationRouter: ObservableObject {
     var path = NavigationPath()  // 네비게이션 경로를 저장하는 변수
 
     /// 특정 화면을 추가 (Push 기능)
-    func push(_ route: Route) {
+    func push(_ route: AppRoute) {
         path.append(route)
     }
