--- weady/weady/Sources/Common/Navigation/Route.swift
@@ -1,25 +1,0 @@
-//
-//  Route.swift
-//  weady
-//
-//  Created by 엄민서 on 7/10/25.
-//
-
-import SwiftUI
-
-enum Route: Hashable {
-    case login
-    case onboarding
-    case basetab
-//    case home
-//    case weadyboard
-//    case weadyboardPost
-////    case weadyboardPostWithItem(WeadyBoardItem)
-////    case weadyboardPostReportDetail(ReportReason)
-////    case weadyboardUpload
-//    case weadychive
-//    case mypage
-//    case weatheraddlocation
-//    case weathersearch
-//    case weatherlocation
-}
